+++ scanpy/api/tl.py
@@ -1,19 +1,19 @@
-from ..tools.pca import pca
-from ..tools.tsne import tsne
-from ..tools.umap import umap
-from ..tools.diffmap import diffmap
-from ..tools.draw_graph import draw_graph
+from ..tools._pca import pca
+from ..tools._tsne import tsne
+from ..tools._umap import umap
+from ..tools._diffmap import diffmap
+from ..tools._draw_graph import draw_graph
 
-from ..tools.paga import paga, paga_degrees, paga_expression_entropies, paga_compare_paths
-from ..tools.rank_genes_groups import rank_genes_groups
-from ..tools.dpt import dpt
-from ..tools.leiden import leiden
-from ..tools.louvain import louvain
-from ..tools.sim import sim
-from ..tools.top_genes import correlation_matrix, ROC_AUC_analysis
+from ..tools._paga import paga, paga_degrees, paga_expression_entropies, paga_compare_paths
+from ..tools._rank_genes_groups import rank_genes_groups
+from ..tools._dpt import dpt
+from ..tools._leiden import leiden
+from ..tools._louvain import louvain
+from ..tools._sim import sim
+from ..tools._top_genes import correlation_matrix, ROC_AUC_analysis
 
-from ..tools.score_genes import score_genes, score_genes_cell_cycle
+from ..tools._score_genes import score_genes, score_genes_cell_cycle
 
-from ..tools.pypairs import cyclone, sandbag
+from ..tools._pypairs import cyclone, sandbag
 
-from ..tools.phate import phate
+from ..tools._phate import phate
