+++ scanpy/tools/__init__.py
@@ -1,13 +1,13 @@
 from ..preprocessing.simple import pca
-from .tsne import tsne
-from .umap import umap
-from .diffmap import diffmap
-from .draw_graph import draw_graph
+from ._tsne import tsne
+from ._umap import umap
+from ._diffmap import diffmap
+from ._draw_graph import draw_graph
 
-from .paga import paga, paga_degrees, paga_expression_entropies, paga_compare_paths
-from .rank_genes_groups import rank_genes_groups
-from .dpt import dpt
-from .leiden import leiden
-from .louvain import louvain
-from .sim import sim
-from .score_genes import score_genes, score_genes_cell_cycle
+from ._paga import paga, paga_degrees, paga_expression_entropies, paga_compare_paths
+from ._rank_genes_groups import rank_genes_groups
+from ._dpt import dpt
+from ._leiden import leiden
+from ._louvain import louvain
+from ._sim import sim
+from ._score_genes import score_genes, score_genes_cell_cycle
