+++ scanpy/tools/_sim.py
